*: add Revision compactor
